% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cookie}
\alias{get_cookie}
\title{Get a cookie value by name from a cookie string}
\usage{
get_cookie(cookie_string, name)
}
\arguments{
\item{cookie_string}{the cookie string}

\item{name}{the name of the cookie}
}
\description{
Get a cookie value by name from a cookie string
}
