% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secure-app.R
\name{secure_server}
\alias{secure_server}
\title{secure shiny server}
\usage{
secure_server(server, ...)
}
\arguments{
\item{server}{server of of application}

\item{...}{not implemented}
}
\description{
secure shiny server
}
