% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verify_token}
\alias{verify_token}
\title{check token is valid or not from server}
\usage{
verify_token(token)
}
\arguments{
\item{token}{token to be verified}
}
\description{
check token is valid or not from server
}
